Extract ProcessInfo, MachineInfo, KillType out from ISimulator
