+++ fdbrpc/include/fdbrpc/AsyncFileNonDurable.actor.h
@@ -45,16 +45,14 @@ ACTOR Future<Void> sendErrorOnProcess(ISimulator::ProcessInfo* process,
                                       Error e,
                                       TaskPriority taskID);
 
+extern Future<Void> waitShutdownSignal();
+
 ACTOR template <class T>
 Future<T> sendErrorOnShutdown(Future<T> in, bool assertOnCancel = false) {
 	try {
 		choose {
-			when(wait(success(g_simulator->getCurrentProcess()->shutdownSignal.getFuture()))) {
-				throw io_error().asInjectedFault();
-			}
-			when(T rep = wait(in)) {
-				return rep;
-			}
+			when(wait(waitShutdownSignal())) { throw io_error().asInjectedFault(); }
+			when(T rep = wait(in)) { return rep; }
 		}
 	} catch (Error& e) {
 		ASSERT(e.code() != error_code_actor_cancelled || !assertOnCancel);
@@ -73,55 +71,17 @@ public:
 		shutdown = doShutdown(this);
 	}
 
-	ACTOR Future<Void> doShutdown(AsyncFileDetachable* self) {
-		wait(success(g_simulator->getCurrentProcess()->shutdownSignal.getFuture()));
-		self->file = Reference<IAsyncFile>();
-		return Void();
-	}
-
-	ACTOR static Future<Reference<IAsyncFile>> open(Future<Reference<IAsyncFile>> wrappedFile) {
-		choose {
-			when(wait(success(g_simulator->getCurrentProcess()->shutdownSignal.getFuture()))) {
-				throw io_error().asInjectedFault();
-			}
-			when(Reference<IAsyncFile> f = wait(wrappedFile)) {
-				return makeReference<AsyncFileDetachable>(f);
-			}
-		}
-	}
+	ACTOR Future<Void> doShutdown(AsyncFileDetachable* self);
+	ACTOR static Future<Reference<IAsyncFile>> open(Future<Reference<IAsyncFile>> wrappedFile);
 
 	void addref() override { ReferenceCounted<AsyncFileDetachable>::addref(); }
 	void delref() override { ReferenceCounted<AsyncFileDetachable>::delref(); }
 
-	Future<int> read(void* data, int length, int64_t offset) override {
-		if (!file.getPtr() || g_simulator->getCurrentProcess()->shutdownSignal.getFuture().isReady())
-			return io_error().asInjectedFault();
-		return sendErrorOnShutdown(file->read(data, length, offset), assertOnReadWriteCancel);
-	}
-
-	Future<Void> write(void const* data, int length, int64_t offset) override {
-		if (!file.getPtr() || g_simulator->getCurrentProcess()->shutdownSignal.getFuture().isReady())
-			return io_error().asInjectedFault();
-		return sendErrorOnShutdown(file->write(data, length, offset), assertOnReadWriteCancel);
-	}
-
-	Future<Void> truncate(int64_t size) override {
-		if (!file.getPtr() || g_simulator->getCurrentProcess()->shutdownSignal.getFuture().isReady())
-			return io_error().asInjectedFault();
-		return sendErrorOnShutdown(file->truncate(size));
-	}
-
-	Future<Void> sync() override {
-		if (!file.getPtr() || g_simulator->getCurrentProcess()->shutdownSignal.getFuture().isReady())
-			return io_error().asInjectedFault();
-		return sendErrorOnShutdown(file->sync());
-	}
-
-	Future<int64_t> size() const override {
-		if (!file.getPtr() || g_simulator->getCurrentProcess()->shutdownSignal.getFuture().isReady())
-			return io_error().asInjectedFault();
-		return sendErrorOnShutdown(file->size());
-	}
+	Future<int> read(void* data, int length, int64_t offset) override;
+	Future<Void> write(void const* data, int length, int64_t offset) override;
+	Future<Void> truncate(int64_t size) override;
+	Future<Void> sync() override;
+	Future<int64_t> size() const override;
 
 	int64_t debugFD() const override {
 		if (!file.getPtr())
@@ -225,58 +185,7 @@ public:
 	                                                std::string actualFilename,
 	                                                Future<Reference<IAsyncFile>> wrappedFile,
 	                                                Reference<DiskParameters> diskParameters,
-	                                                bool aio) {
-		state ISimulator::ProcessInfo* currentProcess = g_simulator->getCurrentProcess();
-		state TaskPriority currentTaskID = g_network->getCurrentTask();
-		state Future<Void> shutdown = success(currentProcess->shutdownSignal.getFuture());
-
-		//TraceEvent("AsyncFileNonDurableOpenBegin").detail("Filename", filename).detail("Addr", g_simulator->getCurrentProcess()->address);
-		wait(g_simulator->onMachine(currentProcess));
-		try {
-			wait(success(wrappedFile) || shutdown);
-
-			if (shutdown.isReady())
-				throw io_error().asInjectedFault();
-
-			state Reference<IAsyncFile> file = wrappedFile.get();
-
-			// If we are in the process of deleting a file, we can't let someone else modify it at the same time.  We
-			// therefore block the creation of new files until deletion is complete
-			state std::map<std::string, Future<Void>>::iterator deletedFile = filesBeingDeleted.find(filename);
-			if (deletedFile != filesBeingDeleted.end()) {
-				//TraceEvent("AsyncFileNonDurableOpenWaitOnDelete1").detail("Filename", filename);
-				wait(deletedFile->second || shutdown);
-				//TraceEvent("AsyncFileNonDurableOpenWaitOnDelete2").detail("Filename", filename);
-				if (shutdown.isReady())
-					throw io_error().asInjectedFault();
-				wait(g_simulator->onProcess(currentProcess, currentTaskID));
-			}
-
-			state Reference<AsyncFileNonDurable> nonDurableFile(
-			    new AsyncFileNonDurable(filename, actualFilename, file, diskParameters, currentProcess->address, aio));
-
-			// Causes the approximateSize member to be set
-			state Future<int64_t> sizeFuture = nonDurableFile->size();
-			wait(success(sizeFuture) || shutdown);
-
-			if (shutdown.isReady())
-				throw io_error().asInjectedFault();
-
-			//TraceEvent("AsyncFileNonDurableOpenComplete").detail("Filename", filename);
-
-			wait(g_simulator->onProcess(currentProcess, currentTaskID));
-
-			return nonDurableFile;
-		} catch (Error& e) {
-			state Error err = e;
-			std::string currentFilename =
-			    (wrappedFile.isReady() && !wrappedFile.isError()) ? wrappedFile.get()->getFilename() : actualFilename;
-			currentProcess->machine->openFiles.erase(currentFilename);
-			//TraceEvent("AsyncFileNonDurableOpenError").errorUnsuppressed(e).detail("Filename", filename).detail("Address", currentProcess->address).detail("Addr", g_simulator->getCurrentProcess()->address);
-			wait(g_simulator->onProcess(currentProcess, currentTaskID));
-			throw err;
-		}
-	}
+	                                                bool aio);
 
 	~AsyncFileNonDurable() override {
 		//TraceEvent("AsyncFileNonDurable_Destroy", id).detail("Filename", filename);
@@ -301,22 +210,7 @@ public:
 	}
 
 	// Removes a file from the openFiles map
-	static void removeOpenFile(std::string filename, AsyncFileNonDurable* file) {
-		auto& openFiles = g_simulator->getCurrentProcess()->machine->openFiles;
-
-		auto iter = openFiles.find(filename);
-
-		// Various actions (e.g. simulated delete) can remove a file from openFiles prematurely, so it may already
-		// be gone. Renamed files (from atomic write and create) will also be present under only one of the two
-		// names.
-		if (iter != openFiles.end()) {
-			// even if the filename exists, it doesn't mean that it references the same file. It could be that the
-			// file was renamed and later a file with the same name was opened.
-			if (iter->second.getPtrIfReady().orDefault(nullptr) == file) {
-				openFiles.erase(iter);
-			}
-		}
-	}
+	static void removeOpenFile(std::string filename, AsyncFileNonDurable* file);
 
 	// Passes along reads straight to the underlying file, waiting for any outstanding changes that could affect the
 	// results
@@ -443,21 +337,7 @@ private:
 		return readFuture.get();
 	}
 
-	ACTOR Future<int> read(AsyncFileNonDurable* self, void* data, int length, int64_t offset) {
-		state ISimulator::ProcessInfo* currentProcess = g_simulator->getCurrentProcess();
-		state TaskPriority currentTaskID = g_network->getCurrentTask();
-		wait(g_simulator->onMachine(currentProcess));
-
-		try {
-			state int rep = wait(self->onRead(self, data, length, offset));
-			wait(g_simulator->onProcess(currentProcess, currentTaskID));
-			return rep;
-		} catch (Error& e) {
-			state Error err = e;
-			wait(g_simulator->onProcess(currentProcess, currentTaskID));
-			throw err;
-		}
-	}
+	ACTOR Future<int> read(AsyncFileNonDurable* self, void* data, int length, int64_t offset);
 
 	// Delays writes a random amount of time before passing them through to the underlying file.
 	// If a kill interrupts the delay, then the output could be the correct write, part of the write,
@@ -511,9 +391,7 @@ private:
 		state bool saveDurable = true;
 		choose {
 			when(wait(delay(delayDuration))) {}
-			when(bool durable = wait(startSyncFuture)) {
-				saveDurable = durable;
-			}
+			when(bool durable = wait(startSyncFuture)) { saveDurable = durable; }
 		}
 
 		debugFileCheck("AsyncFileNonDurableWriteAfterWait", self->filename, dataCopy.begin(), offset, length);
@@ -690,9 +568,7 @@ private:
 		state bool saveDurable = true;
 		choose {
 			when(wait(delay(delayDuration))) {}
-			when(bool durable = wait(startSyncFuture)) {
-				saveDurable = durable;
-			}
+			when(bool durable = wait(startSyncFuture)) { saveDurable = durable; }
 		}
 
 		if (g_network->check_yield(TaskPriority::DefaultYield)) {
@@ -840,50 +716,7 @@ private:
 	}
 
 	// Finishes all outstanding actors on an AsyncFileNonDurable and then deletes it
-	ACTOR Future<Void> closeFile(AsyncFileNonDurable* self) {
-		state ISimulator::ProcessInfo* currentProcess = g_simulator->getCurrentProcess();
-		state TaskPriority currentTaskID = g_network->getCurrentTask();
-		state std::string filename = self->filename;
-
-		g_simulator->getMachineByNetworkAddress(self->openedAddress)
-		    ->deletingOrClosingFiles.insert(self->getFilename());
-
-		wait(g_simulator->onMachine(currentProcess));
-		try {
-			// Make sure all writes have gone through.
-			Promise<bool> startSyncPromise = self->startSyncPromise;
-			self->startSyncPromise = Promise<bool>();
-			startSyncPromise.send(true);
-
-			std::vector<Future<Void>> outstandingModifications;
-
-			for (auto itr = self->pendingModifications.ranges().begin();
-			     itr != self->pendingModifications.ranges().end();
-			     ++itr)
-				if (itr->value().isValid() && !itr->value().isReady())
-					outstandingModifications.push_back(itr->value());
-
-			// Ignore errors here so that all modifications can finish
-			wait(waitForAllReady(outstandingModifications));
-
-			// Make sure we aren't in the process of killing the file
-			if (self->killed.isSet())
-				wait(self->killComplete.getFuture());
-
-			// Remove this file from the filesBeingDeleted map so that new files can be created with this filename
-			g_simulator->getMachineByNetworkAddress(self->openedAddress)->closingFiles.erase(self->getFilename());
-			g_simulator->getMachineByNetworkAddress(self->openedAddress)
-			    ->deletingOrClosingFiles.erase(self->getFilename());
-			AsyncFileNonDurable::filesBeingDeleted.erase(self->filename);
-			//TraceEvent("AsyncFileNonDurable_FinishDelete", self->id).detail("Filename", self->filename);
-
-			delete self;
-			return Void();
-		} catch (Error& e) {
-			state Error err = e;
-			throw err;
-		}
-	}
+	ACTOR Future<Void> closeFile(AsyncFileNonDurable* self);
 };
 
 #include "flow/unactorcompiler.h"
