+++ fdbserver/workloads/ClientWorkload.actor.cpp
@@ -21,6 +21,7 @@
 #include "fdbserver/ServerDBInfo.actor.h"
 #include "fdbserver/workloads/workloads.actor.h"
 #include "fdbrpc/simulator.h"
+#include "fdbrpc/SimulatorProcessInfo.h"
 
 #include <fmt/format.h>
