+++ fdbserver/workloads/DataLossRecovery.actor.cpp
@@ -31,6 +31,7 @@
 #include "flow/Error.h"
 #include "flow/IRandom.h"
 #include "flow/flow.h"
+#include "fdbrpc/SimulatorProcessInfo.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 namespace {
@@ -278,7 +279,7 @@ struct DataLossRecoveryWorkload : TestWorkload {
 	void killProcess(DataLossRecoveryWorkload* self, const NetworkAddress& addr) {
 		ISimulator::ProcessInfo* process = g_simulator->getProcessByAddress(addr);
 		ASSERT(process->addresses.contains(addr));
-		g_simulator->killProcess(process, ISimulator::KillInstantly);
+		g_simulator->killProcess(process, ISimulator::KillType::KillInstantly);
 		TraceEvent("TestTeamKilled").detail("Address", addr);
 	}
