+++ fdbserver/workloads/RandomClogging.actor.cpp
@@ -20,6 +20,7 @@
 
 #include "flow/DeterministicRandom.h"
 #include "fdbrpc/simulator.h"
+#include "fdbrpc/SimulatorProcessInfo.h"
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbserver/TesterInterface.actor.h"
 #include "fdbserver/workloads/workloads.actor.h"
