+++ fdbserver/workloads/RandomMoveKeys.actor.cpp
@@ -211,7 +211,7 @@ struct MoveKeysWorkload : FailureInjectionWorkload {
 	ACTOR Future<Void> forceMasterFailure(Database cx, MoveKeysWorkload* self) {
 		ASSERT(g_network->isSimulated());
 		loop {
-			if (g_simulator->killZone(self->dbInfo->get().master.locality.zoneId(), ISimulator::Reboot, true))
+			if (g_simulator->killZone(self->dbInfo->get().master.locality.zoneId(), ISimulator::KillType::Reboot, true))
 				return Void();
 			wait(delay(1.0));
 		}
